requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt==4.0.1
passlib
pydantic[email]
decouple
alembic